mv error_on_execution_failure to error mod
